Move the vsx code to powerpc
Even if it is unlikely to happen the VSX instructions can be executed in
32bit mode just as well.
